@@ -1,6 +1,6 @@
 module github.com/open-telemetry/opentelemetry-collector-contrib/cmd/schemagen
 
-go 1.24.0
+go 1.25.0
 
 require (
 	github.com/iancoleman/strcase v0.3.0
