@@ -1,6 +1,6 @@
 module github.com/open-telemetry/opentelemetry-collector-contrib/confmap/provider/googlesecretmanagerprovider
 
-go 1.24.0
+go 1.25.0
 
 require (
 	cloud.google.com/go/secretmanager v1.16.0
