@@ -187,14 +187,12 @@ func (s *Stream) run(ctx context.Context, dc doneCancel, streamClient StreamClie
 	var ww sync.WaitGroup
 
 	var writeErr error
-	ww.Add(1)
-	go func() {
-		defer ww.Done()
+	ww.Go(func() {
 		writeErr = s.write(ctx)
 		if writeErr != nil {
 			dc.cancel()
 		}
-	}()
+	})
 
 	// the result from read() is processed after cancel and wait,
 	// so we can set s.client = nil in case of a delayed Unimplemented.
