@@ -194,15 +194,13 @@ func TestConnManagerAcquireReleaseConcurrent(t *testing.T) {
 			test: func(cm *ConnManager) {
 				var wg sync.WaitGroup
 				for range 100 {
-					wg.Add(1)
-					go func() {
-						defer wg.Done()
+					wg.Go(func() {
 						conn, err := cm.Acquire(t.Context())
 						if err != nil {
 							return
 						}
 						cm.Release(t.Context(), conn)
-					}()
+					})
 				}
 				wg.Wait()
 			},
@@ -228,15 +226,13 @@ func TestConnManagerAcquireDiscardConcurrent(t *testing.T) {
 			test: func(cm *ConnManager) {
 				var wg sync.WaitGroup
 				for range 100 {
-					wg.Add(1)
-					go func() {
-						defer wg.Done()
+					wg.Go(func() {
 						conn, err := cm.Acquire(t.Context())
 						if err != nil {
 							return
 						}
 						cm.DiscardAndClose(conn)
-					}()
+					})
 				}
 				wg.Wait()
 			},
