@@ -221,11 +221,9 @@ func newSender(client awsxray.XRayClient, opts ...Option) *telemetrySender {
 // Start starts the loop to send the records.
 func (ts *telemetrySender) Start(ctx context.Context) {
 	ts.startOnce.Do(func() {
-		ts.stopWait.Add(1)
-		go func() {
-			defer ts.stopWait.Done()
+		ts.stopWait.Go(func() {
 			ts.run(ctx)
-		}()
+		})
 	})
 }
 
