@@ -31,9 +31,7 @@ func (i *Input) Start(_ operator.Persister) error {
 	ctx, cancel := context.WithCancel(context.Background())
 	i.cancel = cancel
 
-	i.wg.Add(1)
-	go func() {
-		defer i.wg.Done()
+	i.wg.Go(func() {
 		n := 0
 		for {
 			select {
@@ -57,7 +55,7 @@ func (i *Input) Start(_ operator.Persister) error {
 				return
 			}
 		}
-	}()
+	})
 
 	return nil
 }
