@@ -42,9 +42,7 @@ func (i *Input) Start(_ operator.Persister) error {
 
 	scanner := bufio.NewScanner(i.stdin)
 
-	i.wg.Add(1)
-	go func() {
-		defer i.wg.Done()
+	i.wg.Go(func() {
 		for {
 			select {
 			case <-ctx.Done():
@@ -68,7 +66,7 @@ func (i *Input) Start(_ operator.Persister) error {
 				return
 			}
 		}
-	}()
+	})
 
 	return nil
 }
