@@ -138,11 +138,9 @@ func (p *isolationForestProcessor) Start(_ context.Context, _ component.Host) er
 	// Any additional initialization logic can go here
 
 	// Start the background model update loop
-	p.shutdownWG.Add(1)
-	go func() {
-		defer p.shutdownWG.Done()
+	p.shutdownWG.Go(func() {
 		p.modelUpdateLoop()
-	}()
+	})
 
 	return nil
 }
