@@ -47,13 +47,11 @@ func TestConsumeMetrics(t *testing.T) {
 			idx := processor.cs.add(ch)
 			receiveNum := 0
 			wg := &sync.WaitGroup{}
-			wg.Add(1)
-			go func() {
-				defer wg.Done()
+			wg.Go(func() {
 				for range ch {
 					receiveNum++
 				}
-			}()
+			})
 
 			for i := 0; i < c.limit*2; i++ {
 				// send metric to chan c.limit*2 per sec.
@@ -98,13 +96,11 @@ func TestConsumeLogs(t *testing.T) {
 			idx := processor.cs.add(ch)
 			receiveNum := 0
 			wg := &sync.WaitGroup{}
-			wg.Add(1)
-			go func() {
-				defer wg.Done()
+			wg.Go(func() {
 				for range ch {
 					receiveNum++
 				}
-			}()
+			})
 
 			// send log to chan c.limit*2 per sec.
 			for i := 0; i < c.limit*2; i++ {
@@ -151,13 +147,11 @@ func TestConsumeTraces(t *testing.T) {
 			idx := processor.cs.add(ch)
 			receiveNum := 0
 			wg := &sync.WaitGroup{}
-			wg.Add(1)
-			go func() {
-				defer wg.Done()
+			wg.Go(func() {
 				for range ch {
 					receiveNum++
 				}
-			}()
+			})
 
 			for i := 0; i < c.limit*2; i++ {
 				// send trace to chan c.limit*2 per sec.
