@@ -57,11 +57,9 @@ func Test_Server_ListenAndServe(t *testing.T) {
 			mr.wgMetricsProcessed.Add(1)
 
 			wgListenAndServe := sync.WaitGroup{}
-			wgListenAndServe.Add(1)
-			go func() {
-				defer wgListenAndServe.Done()
+			wgListenAndServe.Go(func() {
 				assert.Error(t, svr.ListenAndServe(p, mc, mr))
-			}()
+			})
 
 			runtime.Gosched()
 
