@@ -118,10 +118,7 @@ func (cfr *cloudFoundryReceiver) Start(ctx context.Context, host component.Host)
 
 	innerCtx, cancel := context.WithCancel(ctx)
 	cfr.cancel = cancel
-	cfr.goroutines.Add(1)
-
-	go func() {
-		defer cfr.goroutines.Done()
+	cfr.goroutines.Go(func() {
 		cfr.settings.Logger.Debug("cloudfoundry receiver starting")
 		_, tokenErr = tokenProvider.ProvideToken()
 		if tokenErr != nil {
@@ -139,7 +136,7 @@ func (cfr *cloudFoundryReceiver) Start(ctx context.Context, host component.Host)
 			cfr.streamMetrics(innerCtx, streamFactory.CreateMetricsStream(innerCtx, cfr.config.RLPGateway.ShardID), host)
 		}
 		cfr.settings.Logger.Debug("cloudfoundry receiver stopped")
-	}()
+	})
 	return nil
 }
 
