@@ -76,13 +76,11 @@ func (cdr *collectdReceiver) Start(ctx context.Context, host component.Host) err
 	if err != nil {
 		return err
 	}
-	cdr.shutdownWG.Add(1)
-	go func() {
-		defer cdr.shutdownWG.Done()
+	cdr.shutdownWG.Go(func() {
 		if err := cdr.server.Serve(l); !errors.Is(err, http.ErrServerClosed) && err != nil {
 			componentstatus.ReportStatus(host, componentstatus.NewFatalErrorEvent(err))
 		}
-	}()
+	})
 	return nil
 }
 
