@@ -128,14 +128,11 @@ func (r *lokiReceiver) startHTTPServer(ctx context.Context, host component.Host)
 	if err != nil {
 		return err
 	}
-	r.shutdownWG.Add(1)
-
-	go func() {
-		defer r.shutdownWG.Done()
+	r.shutdownWG.Go(func() {
 		if errHTTP := r.serverHTTP.Serve(listener); !errors.Is(errHTTP, http.ErrServerClosed) && errHTTP != nil {
 			componentstatus.ReportStatus(host, componentstatus.NewFatalErrorEvent(errHTTP))
 		}
-	}()
+	})
 	return nil
 }
 
@@ -145,14 +142,11 @@ func (r *lokiReceiver) startGRPCServer(ctx context.Context, host component.Host)
 	if err != nil {
 		return err
 	}
-	r.shutdownWG.Add(1)
-
-	go func() {
-		defer r.shutdownWG.Done()
+	r.shutdownWG.Go(func() {
 		if errGRPC := r.serverGRPC.Serve(listener); !errors.Is(errGRPC, grpc.ErrServerStopped) && errGRPC != nil {
 			componentstatus.ReportStatus(host, componentstatus.NewFatalErrorEvent(errGRPC))
 		}
-	}()
+	})
 	return nil
 }
 
