@@ -111,9 +111,7 @@ func (alr *accessLogsReceiver) startPolling(ctx context.Context) error {
 		}
 
 		t := time.NewTicker(pc.AccessLogs.PollInterval)
-		alr.wg.Add(1)
-		go func() {
-			defer alr.wg.Done()
+		alr.wg.Go(func() {
 			for {
 				select {
 				case <-t.C:
@@ -124,7 +122,7 @@ func (alr *accessLogsReceiver) startPolling(ctx context.Context) error {
 					return
 				}
 			}
-		}()
+		})
 	}
 
 	return nil
