@@ -113,9 +113,7 @@ func (er *eventsReceiver) Shutdown(ctx context.Context) error {
 
 func (er *eventsReceiver) startPolling(ctx context.Context) error {
 	t := time.NewTicker(er.pollInterval)
-	er.wg.Add(1)
-	go func() {
-		defer er.wg.Done()
+	er.wg.Go(func() {
 		for {
 			select {
 			case <-t.C:
@@ -126,7 +124,7 @@ func (er *eventsReceiver) startPolling(ctx context.Context) error {
 				return
 			}
 		}
-	}()
+	})
 
 	return nil
 }
